Rework based on feedback from PR
